+++ as-relevansi/blocks/search-block/index.asset.php
@@ -4,7 +4,6 @@ return array(
         'wp-blocks',
         'wp-element',
         'wp-block-editor',
-        'wp-components',
     ),
-    'version' => '0.1.10',
+    'version' => '0.1.11',
 );